refactor : enums 안에 도메인 별로 패키지 생성후 이동 및 경로 적용
