+++ src/main/java/sisosolsol/greenfire/challenge/model/dto/ChallengeCreateDTO.java
@@ -3,7 +3,7 @@ package sisosolsol.greenfire.challenge.model.dto;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-import sisosolsol.greenfire.common.enums.ChallengeStatus;
+import sisosolsol.greenfire.common.enums.challenge.ChallengeStatus;
 
 import java.time.LocalDateTime;
