+++ src/main/java/sisosolsol/greenfire/common/enums/category/CategoryType.java
@@ -1,4 +1,4 @@
-package sisosolsol.greenfire.common.enums;
+package sisosolsol.greenfire.common.enums.category;
 
 public enum CategoryType {
     CHALLENGE,  // 챌린지
