+++ src/main/java/sisosolsol/greenfire/store/model/dto/ApplyStoreListDTO.java
@@ -2,7 +2,7 @@ package sisosolsol.greenfire.store.model.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import sisosolsol.greenfire.common.enums.StoreStatus;
+import sisosolsol.greenfire.common.enums.store.StoreStatus;
 
 import java.time.OffsetDateTime;
